pyproject.toml
[project]
name = "archify-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.136.0",
    "jinja2>=3.1.6",
    "pydantic[email]>=2.13.2",
    "python-dotenv>=1.2.2",
    "supabase>=2.28.3",
    "uvicorn[standard]>=0.44.0",
]
